continued work on preferences page, refresh page after clicking logout will refresh current page for landing page population
