+++ src/components/Burger/Burger.js
@@ -139,6 +139,7 @@ export default function Burger(props) {
   const removeToken = () => {
     localStorage.removeItem("token");
     setToken(null);
+    window.location.reload(true);
   }
 
   return (
@@ -147,7 +148,7 @@ export default function Burger(props) {
         <NavLink exact activeClassName="main-links" to="/watchparty" id='watch-party'>Watch Party</NavLink>
         <NavLink exact activeClassName="main-links" to="/library" id="library">Library</NavLink>
         <NavLink exact activeClassName="main-links" to="/invite" id="invite">Invite</NavLink>
-        <div className="main-links" id="preferences" >Preferences</div>
+        <NavLink exact activeClassName="main-links" to="/preferences" id="preferences">Preferences</NavLink>
         <NavLink exact activeClassName="main-links" to="/" onClick={removeToken} id="logout">Log Out</NavLink>
         {/* <button id="dark-light"onClick={() => props.themeToggler()}>Change Theme</button> */}
